+++ CHANGELOG.md
@@ -3,7 +3,9 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
-- DF-1177 & DF-1161 Added services for GitHub and GitLab with linking to server side scripting.
+### Added
+- DF-1177 & DF-1161 Added services for GitHub and GitLab with linking to server side scripting
+- Support for setting df.scripting.default_protocol in config and DF_SCRIPTING_DEFAULT_PROTOCOL in .env
 
 ## [0.6.0] - 2017-08-17
 ### Changed