+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## [0.6.1] - 2017-09-18
 ### Added
 - DF-1177 & DF-1161 Added services for GitHub and GitLab with linking to server side scripting
 - Support for setting df.scripting.default_protocol in config and DF_SCRIPTING_DEFAULT_PROTOCOL in .env
@@ -52,7 +54,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## 0.1.0 - 2016-11-30
 First official release working with the new [dreamfactory](https://github.com/dreamfactorysoftware/dreamfactory) project.
 
-[Unreleased]: https://github.com/dreamfactorysoftware/df-script/compare/0.6.0...HEAD
+[Unreleased]: https://github.com/dreamfactorysoftware/df-script/compare/0.6.1...HEAD
+[0.6.1]: https://github.com/dreamfactorysoftware/df-script/compare/0.6.0...0.6.1
 [0.6.0]: https://github.com/dreamfactorysoftware/df-script/compare/0.5.0...0.6.0
 [0.5.0]: https://github.com/dreamfactorysoftware/df-script/compare/0.4.0...0.5.0
 [0.4.0]: https://github.com/dreamfactorysoftware/df-script/compare/0.3.0...0.4.0